chore: move syslog package in pkg/util/log
