+++ pkg/util/log/syslog/syslog_test.go
